wip, still no editor showing, moved js and css to own files
